DT-124
Create a connection to the database through Node JS/Express. Note that in order to run and gain access through the database, you need to go and run cd into the Database directory and do 'node DatabaseConnection.tsx' on one terminal, once the console says 'up', it is up and running. Leave that terminal up and open another one for expo. For axios, you might need to change the IP address to the host of the computer running express. An example how to use it is in ModifyAv, and will return a string version of array. Besides customQuery, there are general queries for each table/view, look into DatabaseConnection.tsx for more details.