DT-111 and DT-92 (Calendar selection functionality)
Updated the calendar to accept selection, and store selected dates in an array. Be warned, I had to adjust some formatting in the CSS files since otherwise the calendar had issues with sizing. There may need to be additional readjustments to display properly when adding additional info.

Since Adrian and I have such similar tasks we worked on them together, since once you select an item it's very easy to store it.

Co-Authored-By: VoicelessBark44 <34280572+VoicelessBark44@users.noreply.github.com>