DT 134 Update (not fully functional)
Had to do a lot of digging to figure out how to get useRef working. Still not able to send the dates to the next page though. I think the dates will have to be rendered outside of the component, hence the renderSelectedDates() function. Still didn't work but I'm going to take another crack at it soon. Just wanted to post this in case someone else knew a bit better than I what to do.