DT-24 Adds new query For finding UserID from email
Adds a new Query to the CurrentClientView to find the userID based on the email provided. email needs to be specific though. Will need to be improved in the future. After finding the userid if it exists it updates privs based on situation and then sends that info when autonavigates to homescreen