DT-66 Edit Button
Currently set by default to dummy data, can be edited. The bottom info box titled "Preferences" looks a bit buggy when the edit button is pressed. No issue with the "Customer Info" edit button, so I think it may be a styling issue or something like that, because the text box does change size slightly.
Also, if multiple lines need to be entered they are treated as one input, not divided into an array or anything. Assuming that's not a problem or not too hard to fix
No database connection